Add CSG utils
